completion: implement symbol extractor
SymbolExtractor is a parser for member accesses that the Vala parser
fails to detect. It works in three steps:

1. reconstruct the member access
2. set the first member to the corresponding symbol in the symbol table
   of the current scope (if we can find the current scope)
3. use the Vala semantic analyzer to determine the symbol that the
   entire member access refers to, whether it should be visible, etc

Note that (2) can fail if the current block itself fails to parse (e.g,
because of an invalid member access at the end of the block). In that
case however, we fallback to checking the first valid ancestor block,
which might be the CodeContext root. Therefore, this behavior would be
improved with additional improvements to the Vala parser's error
recovery.

Even still, in many cases where the parsing of blocks is not affected
(say, a member access before a local variable declaration, or before a
return statement), this approach works.