find_scope: search for lambdas in local variables
It's possible for lambda expressions to be embedded in the initializer
of a local temporary variable that was created automatically. Usually
this happens when the compiler converts an expression into multiple
temporaries, such as an expression that could throw an error. For
example, this is the case for the regex calls in
DocComment.from_valadoc_comment() in doccomment.vala.