implement snippets for method completion
When a method is selected, create a snippet to fill in each argument and
type parameter of the method. Do the same thing when selecting an
instantiable type while inside an ObjectCreationExpression.

Additionally, run a style analyzer in order to generate the snippet
based on the code style used. Some codebases use one space before the
method arguments while others use zero spaces.