defaultproject: handle opening managed file
Handle attempts to open a file that is part of a default project.
Although default projects are for opening standalone files, they include
multiple VAPIs (glib.vapi, gobject.vapi). If the user opens a standalone
file, then tries to open and close a VAPI, this would close the default
project, which is not what we want.